pyproject.toml
[project]
name = "OpenRecon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "builtwith>=1.3.4",
    "dnspython>=2.7.0",
    "python-whois>=0.9.5",
    "requests>=2.32.5",
    "sockets>=1.0.0",
    "tldextract>=5.3.0",
]
